requirements.txt
anyio==3.6.2
APScheduler==3.9.1.post1
autopep8==2.0.1
backports.zoneinfo==0.2.1
cachetools==5.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.4
lxml==4.9.1
pkg-resources==0.0.0
psutil==5.9.4
pycodestyle==2.10.0
python-qbittorrent==0.4.3
python-telegram-bot==20.0a0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tornado==6.1
tpblite==0.8.0
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
